Use IConfigureOptions<T> to configure ForwardedHeadersOptions
