+++ WebApps/Proxy/Microsoft.SPID.Proxy/appsettings.json
@@ -159,6 +159,7 @@
 		"CacheAbsoluteExpirationInMins": 120
 	},
 	"ForwardedHeaders": {
+		//"AllowedHosts": [ "allowedhost1.com", "allowedhost2.com", "allowedhostN.com" ],
 		"ForwardedHeaders": "All"
 	}
 }
